requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
pydantic>=2.7
python-dotenv>=1.0

# LangChain + OpenAI split packages
langchain>=0.2
langchain-openai>=0.1.7
openai>=1.45

# Vector DB
chromadb>=0.5

# PDF parsing
PyPDF2>=3.0

# Tokenization (optional but helpful)
tiktoken>=0.7

# Frontend/UI
streamlit>=1.36
requests>=2.31
